feat(test): Allow overriding plugin upload setting for test env
Adds the MM_PLUGINSETTINGS_ENABLEUPLOADS environment variable to the docker-compose.harden.yml file. This allows enabling or disabling plugin uploads for the test environment via the .env.test file.

This is necessary because this setting is locked down in the Mattermost API for security reasons and cannot be changed with mmctl.